+++ src/value.rs
@@ -1,6 +1,6 @@
 #[derive(Debug, Clone)]
 pub struct Value {
-    pub value: u8,
+    pub value: u64,
     pub first_col: usize,
     pub second_col: usize,
     pub col_sum: usize,
